+++ CHANGELOG.md
@@ -118,6 +118,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1623](https://github.com/regen-network/regen-ledger/pull/1623) Add `Msg/UpdateBatchMetadata`
 - [#1625](https://github.com/regen-network/regen-ledger/pull/1625) Add `reason`/`retirement_reason`
 - [#1634](https://github.com/regen-network/regen-ledger/pull/1634) Add `balances-by-batch` query command
+- [#1636](https://github.com/regen-network/regen-ledger/pull/1636) Add `bridge` transaction command
 
 #### Changed